Fix Ttk_CreateTagTable() signature

+++ generic/ttk/ttkWidget.h
@@ -214,7 +214,7 @@ typedef struct TtkTagSet {	/* TODO: make opaque */
 } *Ttk_TagSet;
 
 MODULE_SCOPE Ttk_TagTable Ttk_CreateTagTable(
-	Tcl_Interp *, Tk_Window tkwin, const Tk_OptionSpec *, int recordSize);
+	Tcl_Interp *, Tk_Window tkwin, const Tk_OptionSpec *, size_t recordSize);
 MODULE_SCOPE void Ttk_DeleteTagTable(Ttk_TagTable);
 
 MODULE_SCOPE Ttk_Tag Ttk_GetTag(Ttk_TagTable, const char *tagName);
